qa: extract external image host policy and add tests
- Move external image host validation out of ArticulateImportMediaService into a new ExternalImageHostPolicy class (host normalization, always-blocked hosts/suffixes, IP/address checks and NAT64 handling).
- Update ArticulateImportMediaService and BlogMlApiController to call ExternalImageHostPolicy methods and remove the in-file validation helpers.
Add unit tests for the new policy (ExternalImageHostPolicyTests), adjust ArticulateImportMediaService tests (rename one test and add ArticulateOptions to CreateSut), and apply a small README wording tweak about blocked hosts.
